requirements.txt
feedparser>=6.0.10
requests>=2.31.0
trafilatura>=1.6.0
beautifulsoup4>=4.12.0
python-dateutil>=2.8.2
google-generativeai>=0.3.0
python-dotenv>=1.0.0
